Remove comments components from web harmony
