--- packages/harmony/src/components/comments/ArtistPick/index.ts
@@ -1,2 +1,0 @@
-export { ArtistPick } from './ArtistPick'
-export * from './types'
