--- packages/harmony/src/components/comments/ArtistPick/types.ts
@@ -1,4 +1,0 @@
-export type ArtistPickProps = {
-  isPinned?: boolean
-  isLiked?: boolean
-}
