--- packages/harmony/src/components/comments/CommentText/index.ts
@@ -1,2 +1,0 @@
-export { CommentText } from './CommentText'
-export * from './types'
