--- packages/harmony/src/components/comments/CommentText/types.ts
@@ -1,4 +1,0 @@
-export type CommentTextProps = {
-  children: string
-  isEdited?: boolean
-}
