--- packages/harmony/src/components/comments/Identifier/Identifier.stories.tsx
@@ -1,24 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react'
-
-import { Flex } from 'components/layout'
-
-import { Identifier } from './Identifier'
-
-const meta: Meta<typeof Identifier> = {
-  title: 'Components/Comments/Identifier [beta]',
-  component: Identifier
-}
-
-export default meta
-
-type Story = StoryObj<typeof Identifier>
-
-export const Default: Story = {
-  render: () => (
-    <Flex direction='column' gap='l' p='s'>
-      <Identifier type='artist' />
-      <Identifier type='supporter' />
-      <Identifier type='topSupporter' />
-    </Flex>
-  )
-}
