--- packages/harmony/src/components/comments/Identifier/index.ts
@@ -1,2 +1,0 @@
-export { Identifier } from './Identifier'
-export * from './types'
