--- packages/harmony/src/components/comments/Identifier/types.ts
@@ -1,5 +1,0 @@
-export type IdentifierType = 'artist' | 'topSupporter' | 'supporter'
-
-export type IdentifierProps = {
-  type: IdentifierType
-}
