--- packages/harmony/src/components/comments/Timestamp/index.ts
@@ -1,3 +1,0 @@
-export { Timestamp } from './Timestamp'
-export { getLargestTimeUnitText } from './util'
-export * from './types'
