+++ packages/harmony/src/components/icon-text/IconText.tsx
