+++ packages/harmony/src/components/icon-text/index.ts
