+++ packages/harmony/src/components/index.ts
@@ -25,10 +25,6 @@ export { default as LoadingSpinner } from './loading-spinner/LoadingSpinner'
 export * from './pill'
 export * from './common/HiddenInput'
 export * from './select'
-export * from './comments/ArtistPick'
-export * from './comments/CommentText'
-export * from './comments/IconText'
-export * from './comments/Identifier'
-export * from './comments/SendIcon'
-export * from './comments/Timestamp'
+export * from './icon-text'
+export * from './send-icon'
 export * from './checkbox'
